简化头文件
