+++ .gitignore
@@ -1,4 +1,3 @@
 cmake-build-release
 cmake-build-debug
 .idea
-.gitignore
