+++ src/steamIconFix.cpp
@@ -3,7 +3,7 @@
 //
 
 #include "steamIconFix.h"
-#include "tools.h"
+
 BOOL isFileExists(const string& path) {
     DWORD dwAttr = GetFileAttributes((LPCSTR)path.c_str());
     if (dwAttr == 0xFFFFFFFF) return false;
@@ -118,5 +118,19 @@ bool getStartMenuProgramDir(string &path) {
     }
     *path = wstring2string(wstrPath)+"\\Steam";
     return true;
+}
+string wstring2string(wstring wstr) {
+    string result;
+    int len = WideCharToMultiByte(CP_ACP,WC_NO_BEST_FIT_CHARS,
+                                  wstr.c_str(),wstr.size(),
+                                  nullptr,0, nullptr, nullptr);
+    char* buffer=new char [len+1];
+    WideCharToMultiByte(CP_ACP,WC_NO_BEST_FIT_CHARS,
+                        wstr.c_str(),wstr.size(),
+                        buffer,len, nullptr, nullptr);
+    buffer[len]='\0';
+    result.append(buffer);
+    delete[] buffer;
+    return result;
 }
