--- src/tools.h
@@ -1,14 +1,0 @@
-//
-// Created by -Zinc- on 2024/3/10.
-//
-
-#ifndef STEAMICONFIX_TOOLS_H
-#define STEAMICONFIX_TOOLS_H
-
-#include <string>
-#include <winsock2.h>
-#include <windows.h>
-
-using namespace std;
-string wstring2string(wstring wstr);
-#endif //STEAMICONFIX_TOOLS_H
